+++ Task1.8/IO/ConsoleWriter.cs
@@ -4,7 +4,7 @@ public static class ConsoleWriter
 {
     private const char NumberSeparator = ' ';
     
-    public static void WriteLn( IEnumerable<float> numbers )
+    public static void WriteLn( IEnumerable<double> numbers )
     {
         Console.WriteLine( string.Join( NumberSeparator, numbers ) );
     }
